Postal Service program
You have recently started working for the postal service and your job is to make sure that the post is delivered as efficiently as possible. Each
morning a postal worker gives you a text file containing the distance between the houses that he needs to visit.
The ordering of the entries in the text file can be in any order as long as each road is denoted by {Point A, Point B, Distance}.
It is possible that the same entry might occur more than once in the text file. If they have different distances the bigger distance would be
the correct distance.
The movement between two points is bidirectional, eg. {A, B, 1} and {B, A, 1} are both the same.
The file might contain whitespace and be on multiple lines.
All entries will be complete so you don't need to worry about something like {A, B
An example text file has the following format
{A,B,3} {A,C,2}{A,C, 2}
{A,D,3}{B, C,1}
In this instance the postal worker would need to travel as follows, A->B->C->A->D with a travel length of 9.